refactorHide Micrometer CircuitBreaker metrics behind trait, generalize http4s-client-monix-catnap module
