+++ monix-catnap-micrometer/src/main/scala/com/avast/sst/monix/catnap/micrometer/MicrometerCircuitBreakerMetrics.scala
@@ -5,32 +5,37 @@ import java.util.concurrent.atomic.AtomicInteger
 import cats.effect.Sync
 import com.avast.sst.monix.catnap.CircuitBreakerMetrics
 import com.avast.sst.monix.catnap.CircuitBreakerMetrics.State
-import com.avast.sst.monix.catnap.CircuitBreakerMetrics.State._
+import com.avast.sst.monix.catnap.CircuitBreakerMetrics.State.{Closed, HalfOpen, Open}
 import io.micrometer.core.instrument.MeterRegistry
 
-/** Implements [[com.avast.sst.monix.catnap.CircuitBreakerMetrics]] for any Micrometer [[io.micrometer.core.instrument.MeterRegistry]]. */
-class MicrometerCircuitBreakerMetrics[F[_]: Sync](name: String, meterRegistry: MeterRegistry) extends CircuitBreakerMetrics[F] {
+trait MicrometerCircuitBreakerMetrics[F[_]] extends CircuitBreakerMetrics[F]
 
-  private val F = Sync[F]
+object MicrometerCircuitBreakerMetrics {
 
-  private[this] val CircuitOpened = -1
-  private[this] val CircuitHalfOpened = 0
-  private[this] val CircuitClosed = 1
+  private[micrometer] class Impl[F[_]: Sync](name: String, meterRegistry: MeterRegistry) extends MicrometerCircuitBreakerMetrics[F] {
 
-  private val accepted = meterRegistry.counter(s"circuit-breaker.$name.accepted")
-  private val rejected = meterRegistry.counter(s"circuit-breaker.$name.rejected")
-  private val circuitState = meterRegistry.gauge[AtomicInteger](s"circuit-breaker.$name.state", new AtomicInteger(CircuitClosed))
+    private val F = Sync[F]
 
-  override def increaseAccepted: F[Unit] = F.delay(accepted.increment())
+    private[this] val CircuitOpened = -1
+    private[this] val CircuitHalfOpened = 0
+    private[this] val CircuitClosed = 1
 
-  override def increaseRejected: F[Unit] = F.delay(rejected.increment())
+    private val accepted = meterRegistry.counter(s"circuit-breaker.$name.accepted")
+    private val rejected = meterRegistry.counter(s"circuit-breaker.$name.rejected")
+    private val circuitState = meterRegistry.gauge[AtomicInteger](s"circuit-breaker.$name.state", new AtomicInteger(CircuitClosed))
 
-  override def setState(state: State): F[Unit] = {
-    state match {
-      case Closed   => F.delay(circuitState.set(CircuitClosed))
-      case Open     => F.delay(circuitState.set(CircuitOpened))
-      case HalfOpen => F.delay(circuitState.set(CircuitHalfOpened))
+    override def increaseAccepted: F[Unit] = F.delay(accepted.increment())
+
+    override def increaseRejected: F[Unit] = F.delay(rejected.increment())
+
+    override def setState(state: State): F[Unit] = {
+      state match {
+        case Closed   => F.delay(circuitState.set(CircuitClosed))
+        case Open     => F.delay(circuitState.set(CircuitOpened))
+        case HalfOpen => F.delay(circuitState.set(CircuitHalfOpened))
+      }
     }
+
   }
 
 }
